SDSTOR-7548 : Fixing std::bad_cast (#1083)
link: https://jirap.corp.ebay.com/browse/SDSTOR-7548

In the test_log_store, there is an inconsistency in SISL_OPTION['device_list'] where it is set as a default value of string and called as the vector of strings. This results in throwing core dump as bad_cast. Using this change fixes the problem as the following example (bin/test_log_store --iterations=10  --device_list traddr=0000:3b:00.0):

(gdb) p m_dev_names
(gdb)$1 = std::vector of length 1, capacity 1 = {"traddr=0000:3b:00.0"}