Persistence of allocated blocks during CP and moved to C++-20.
Without this commit, homestore 4.x doesn't persist the bitmap of allocated blks during CP.
This commit handles that gap. On top of that, the free blk id collection is moved to VirtualDev
layer, so that both index and data service has common blk id collection. This blkid collection
is essential so that all free blk ids are moved to the next CP after the current CP is flushed.

Changed it to C++-20 standard which resulted in few compilation errors, fixed that.