Fix varsize allocator
previously if alloc with is_contiguous = true, we will not try other
portion if current portion do not have space. This is wrong as
contiguous was ensured by `get_next_contiguous_n_reset_bits`.

This commit fixed the behavior as well as improving logging correctness,
before the fix we un-conditionally change a new portion and log a
misleading log "alloc direct unable to find in prev portion" even when
alloction is success.

MetaBlkMgrTest::recovery_test is a victim of this bug as it tries to use
up all the space of meta vdev and will easily hit the case of one
portion full.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>