Fix varsize allocator (#440)
* Fix varsize allocator

previously if alloc with is_contiguous = true, we will not try other
portion if current portion do not have space. This is wrong as
contiguous was ensured by `get_next_contiguous_n_reset_bits`.

This commit fixed the behavior as well as improving logging correctness,
before the fix we un-conditionally change a new portion and log a
misleading log "alloc direct unable to find in prev portion" even when
alloction is success.

MetaBlkMgrTest::recovery_test is a victim of this bug as it tries to use
up all the space of meta vdev and will easily hit the case of one
portion full.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>

* Save m_start_portion_num to local var to avoid racing

m_start_portion_num will be updated for every successful alloc.

it is possible it get updated when another thread is searching for
available portion, which will end the search.

Thread A:

[06/11/24 17:53:28-07:00] [trace] [2618720] [varsize_blk_allocator.cpp:620]
 [blkalloc=] alloc direct unable to find in curr portion 10, will searching
 in portion=11, start_portion=10,continue=true, out_blkid num_pieces=0 , max_pieces=1

[06/11/24 17:53:28-07:00] [debug] [2618720] [varsize_blk_allocator.cpp:609]
 [blkalloc=] Allocated directly from portion=11 nnblks=221 Blk_num=180224 nblks=221 set_bit_count=16291

Thread B:

[06/11/24 17:53:28-07:00] [trace] [2618716] [varsize_blk_allocator.cpp:620]
 [blkalloc=] alloc direct unable to find in curr portion 10, will searching
 in portion=11, start_portion=11,continue=true, out_blkid num_pieces=0 , max_pieces=1

Thread A/B were both searching from portion 10, both failed. Thread A
continue to portion 11 and succeed,  then update the
m_start_portion_num = 11.  Thread B when it about to try 11, it saw
portion(11) == start_portion(11, updated by thread A) , mistakenly given
up.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>

---------

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>