fix ut.
Previously commit dont flush the over-written log,  and the log_entries
only look at flushed logs. As a result, the over-written log was
not included in the log_entries() result which breaks the UT.

The UT was running in master, because we dont at the over_written
entry into cache and the entry_at() call will inline flush this entry.
After adding it to cache it served from cache without flushing.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>