Implement get_next_batch_size_hint_in_bytes()
we use the `byte` as `cnt` as of now.

Also update the log_entries_ext() which will be called on leader,

if hint < 0 means follower want nothing, return an empty vector
so that an empty append_entries_req will be sent, to carry the
commit_index update and trigger the follower to commit.

if hint > 0, respect the cnt that the follower want,  this is useful
when two logs within same batch has dependency, we can exclude the
dependent one.

if hint = 0 means control by leader.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>