fix potential bug of home raft log store initialization
the home raft log store set the log_found_cb and log_replay_done_cb in
the result future of open_log_store, which is an async call. so there is
a chance that when repl_dev start replaying log , the above two callback
are not registered yes, which might lead to some error. this PR move the
register out of the future to the initialization phase of log store,
which will avoid this case