Initializing buffer while creating superblk
This ensures that any uninitialized memory does not affect the data being written or loaded.
For example, if a variable in superblk is `char name[50]` and the real name length is 10,
uninitialized memory may cause the name read from the superblk to be longer than 10 characters.

This PR also update nuraft_mesg version which has https://github.com/eBay/nuraft_mesg/pull/117