Send response if failed to save pushed_data
The previous implementation missed sending a response when storing pushed_data failed.
As a result, a large amount of RPC data was held in memory and only released when the connection timed out,
leading to increased memory usage. This change fixes the issue.

The issue occurs in test case `full pg recovery in 5 replicas env, blob_size=512KB`,
when a batch of old shards sealed and a batch of new shards created, a lot of 'pushed_data' failed to save due to shards not committed.
Then the memory increased and the pod is OOM in the test.